requirements.txt
fastapi
uvicorn[standard]
google-cloud-compute
pydantic
openai
boto3
